requirements.txt
pandas
geopandas
shapely
osmnx
networkx
psycopg2-binary
plotly
scikit-learn
statsmodels
python-dotenv
